Promotion Search
Added Promotion search

+++ API/PromotionApi/Models/Promotion.cs
@@ -7,7 +7,7 @@ namespace PromotionApi.Models
     public class Promotion
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [MaxLength(45)]
         public string Name { get; set; }
         public double Price {get; set; }
@@ -17,7 +17,7 @@ namespace PromotionApi.Models
         public string ImageUrl { get; set; }
         [ForeignKey("StoreFK")]
         public Store Store { get; set; }
-        public int StoreFK { get; set; }
+        public long StoreFK { get; set; }
         [ForeignKey("UserFK")]
         public User User { get; set; }
         public long UserFK { get; set; }
